[dynamo] Overwrite proxy of subgraph outputs
This ensures that
1) We use the same VaribleTracker for a fake tensor value. Earlier, one
   fake value could be mapped to multiple VTs.
2) If we are trying to construct a VT for the same fake tensor, we
   overwrite the proxy of the TensorVariable.

This ensures that if a TensorVariable was mutated in speculate_subgraph,
we make it available from the subgraph output.

This DOES not solve the overall problem of tensors missed in the side
effect. It STILL relies on the user to make the side-effected return
from the subgraph.

ghstack-source-id: 6ab09e4c820b0a5e99cef12182086c5c08df034c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/167029